refactor: Move konnector policies to their own directory
