+++ packages/cozy-harvest-lib/src/policies/clisk.spec.js
@@ -1,4 +1,4 @@
-import { getLauncher, konnectorPolicy } from './cccPolicy'
+import { getLauncher, konnectorPolicy } from './clisk'
 
 describe('getLauncher', () => {
   it('should get the current Launcher from the given window object', () => {
